backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
google-generativeai==0.5.0

# Web scraping dependencies
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
feedparser==6.0.10
